refactor: move function
The `tryResolvePackumentVersion` now only really acts on `UnityPackument` and thus can be moved into the corresponding domain module.
