+++ src/cli/cmd-add.ts
@@ -31,7 +31,7 @@ import { CmdOptions } from "./options";
 import {
   PackumentResolveError,
   pickMostFixable,
-  VersionNotFoundError,
+  
 } from "../packument-resolving";
 import { SemanticVersion } from "../domain/semantic-version";
 import { areArraysEqual } from "../utils/array-utils";
@@ -52,6 +52,7 @@ import { logValidDependency } from "./dependency-logging";
 import { unityRegistryUrl } from "../domain/registry-url";
 import { tryGetTargetEditorVersionFor } from "../domain/package-manifest";
 import { FetchPackumentError } from "../services/fetch-packument";
+import {VersionNotFoundError} from "../domain/packument";
 
 export class InvalidPackumentDataError extends CustomError {
   private readonly _class = "InvalidPackumentDataError";
