+++ src/cli/error-logging.ts
@@ -6,10 +6,11 @@ import { NotFoundError } from "../io/file-io";
 import { Logger } from "npmlog";
 import {
   PackumentResolveError,
-  VersionNotFoundError,
+  
 } from "../packument-resolving";
 import { PackumentNotFoundError } from "../common-errors";
 import { DomainName } from "../domain/domain-name";
+import {VersionNotFoundError} from "../domain/packument";
 
 /**
  * Logs a {@link ManifestLoadError} to the console.
