+++ src/domain/packument.ts
@@ -3,6 +3,10 @@ import { DomainName } from "./domain-name";
 import { UnityPackageManifest } from "./package-manifest";
 import { PackageId } from "./package-id";
 import { Dist, Maintainer } from "@npm/types";
+import { CustomError } from "ts-custom-error";
+import { Err, Ok, Result } from "ts-results-es";
+import { recordKeys } from "../utils/record-utils";
+import { ResolvableVersion } from "../packument-resolving";
 
 /**
  * Contains information about a specific version of a package. This is based on
@@ -115,4 +119,66 @@ export function tryGetPackumentVersion(
   version: SemanticVersion
 ): UnityPackumentVersion | null {
   return packument.versions[version] ?? null;
+}
+
+/**
+ * Error for when a packument with the searched name was found, but it
+ * had no versions.
+ */
+export class NoVersionsError extends CustomError {
+  private readonly _class = "NoVersionsError";
+
+  constructor() {
+    super("A packument contained no versions");
+  }
+}
+
+/**
+ * Error for when a packument with the searched name was found, but a specific
+ * requested version did not exist.
+ */
+export class VersionNotFoundError extends CustomError {
+  private readonly _class = "VersionNotFoundError";
+
+  constructor(
+    /**
+     * The version that was requested.
+     */
+    readonly requestedVersion: SemanticVersion,
+    /**
+     * A list of available versions.
+     */
+    readonly availableVersions: ReadonlyArray<SemanticVersion>
+  ) {
+    super("The requested version was not in the packument.");
+  }
+}
+
+/**
+ * Attempts to resolve a specific version from a packument. Resolve here means
+ * that this function also accepts non-semantic versions and will attempt to
+ * find the correct version from the packument. This differentiates it from
+ * {@link tryGetPackumentVersion}.
+ * @param packument The packument to search.
+ * @param requestedVersion The requested version.
+ */
+export function tryResolvePackumentVersion(
+  packument: UnityPackument,
+  requestedVersion: ResolvableVersion
+): Result<UnityPackumentVersion, NoVersionsError | VersionNotFoundError> {
+  const availableVersions = recordKeys(packument.versions);
+  if (availableVersions.length === 0) return Err(new NoVersionsError());
+
+  // Find the latest version
+  if (requestedVersion === undefined || requestedVersion === "latest") {
+    let latestVersion = tryGetLatestVersion(packument);
+    if (latestVersion === undefined) latestVersion = availableVersions.at(-1)!;
+    return Ok(tryGetPackumentVersion(packument, latestVersion)!);
+  }
+
+  // Find a specific version
+  if (!availableVersions.includes(requestedVersion))
+    return Err(new VersionNotFoundError(requestedVersion, availableVersions));
+
+  return Ok(tryGetPackumentVersion(packument, requestedVersion)!);
 }
