+++ src/services/resolve-remote-packument.ts
@@ -5,10 +5,11 @@ import {
   PackumentResolveError,
   ResolvableVersion,
   ResolvedPackument,
-  tryResolvePackumentVersion,
+  
 } from "../packument-resolving";
 import { PackumentNotFoundError } from "../common-errors";
 import { FetchPackumentError, FetchPackumentService } from "./fetch-packument";
+import {tryResolvePackumentVersion} from "../domain/packument";
 
 /**
  * Service function for resolving remove packuments.
