+++ test/cli/cmd-deps.test.ts
@@ -9,9 +9,9 @@ import { makePackageReference } from "../../src/domain/package-reference";
 import { makeMockLogger } from "./log.mock";
 import { makeSemanticVersion } from "../../src/domain/semantic-version";
 import { PackumentNotFoundError } from "../../src/common-errors";
-import { VersionNotFoundError } from "../../src/packument-resolving";
 import { ResolveDependenciesService } from "../../src/services/dependency-resolving";
 import { mockService } from "../services/service.mock";
+import {VersionNotFoundError} from "../../src/domain/packument";
 
 const somePackage = makeDomainName("com.some.package");
 const otherPackage = makeDomainName("com.other.package");
