+++ test/services/resolve-latest-version.test.ts
@@ -4,10 +4,9 @@ import { makeResolveLatestVersionService } from "../../src/services/resolve-late
 import { makeDomainName } from "../../src/domain/domain-name";
 import { PackumentNotFoundError } from "../../src/common-errors";
 import { Ok } from "ts-results-es";
-import { UnityPackument } from "../../src/domain/packument";
+import {NoVersionsError, UnityPackument} from "../../src/domain/packument";
 import { Registry } from "../../src/domain/registry";
 import { exampleRegistryUrl } from "../domain/data-registry";
-import { NoVersionsError } from "../../src/packument-resolving";
 import { unityRegistryUrl } from "../../src/domain/registry-url";
 
 describe("resolve latest version service", () => {
